--- a/README
+++ b/README
@@ -23,7 +23,7 @@ are contained.
 To build:
 
 export TARGET_OSSL_INCLUDE_PATH=/path/to/openssl/include
-export TARGET_OSSL_LIBRARY_PATH=/path/to/openssl
+export TARGET_OSSL_LIBRARY_PATH=/path/to/openssl/lib
 make
 
 The performance testing apps must be run ensuring that libcrypto.so and
@@ -31,7 +31,7 @@ libssl.so are on the library path.
 
 For example:
 
-LD_LIBRARY_PATH=/path/to/openssl ./randbytes 10
+LD_LIBRARY_PATH=/path/to/openssl/lib ./randbytes 10
 
 Each performance testing app will take different parameters. They are described
 individually below. All performance testing apps take the "--terse" option
@@ -125,9 +125,21 @@ The pkeyread test repeatedly calls the PEM_read_bio_PrivateKey() function on a
 memory BIO with a private key of desired type, when it is running in pem mode
 (-f pem).  If test is running in der mode (-f der) it calls to
 d2i_PrivateKey_ex() function to repeatedly read private key of desired type.
-It does 100000 repetitions divided evenly among each thread. The number of
+It does 10000 repetitions divided evenly among each thread. The number of
 threads to use is provided by option -t. The test reports average time per
 call. Use option -k to select key type for benchmark.  The list of keys for
-testing is as follows: dh, dhx, dsa, ec, rsa, xkey.  To run bencmark for all
+testing is as follows: dh, dhx, dsa, ec, rsa, xkey.  To run benchmark for all
 keys and formats using 4 threads run pkeyread as follows:
     ./pkeyread -f all -k all -t 4
+
+evp_setpeer
+-------
+
+The evp_setpeer test repeatedly calls the EVP_PKEY_derive_set_peer() function
+on a memory BIO with a private key of desired type.  It does 10000
+repetitions divided evenly among each thread. The number of threads to use is
+provided by option -t. The test reports average time per call. Use option -k
+to select key type for benchmark.  The list of keys for testing is as follows:
+dh, ec256, ec521, x25519.  To run benchmark for all keys using 4 threads, run
+evp_setpeer as follows:
+    ./evp_setpeer -k all -t 4