… source/genkeys_setpeer.sh
… source/keys_setpeer.h