Adds a performance test profiling the run time of `EVP_PKEY_derive_set_peer`. This test is very similar to how pkeyread perf test is set up and run.

I created a new `source/genkeys_setpeer.sh` file that can be used to codegen a `keys_setpeer.h` header file with private keys for the test to use. The test will go through the key types (specified as a CLI option), runs `EVP_PKEY_derive_set_peer` 10k times split between however many threads is specified, and prints the results for each key type.

For now, it measures the performance on DH ffdhe2048, ECDH P-256, P-521, and X25519.

OpenSSL master branch:
```console
$ ./evp_setpeer -k all 2
Average time per dh evp_set_peer call: 4191.861000us
Average time per ec256 evp_set_peer call: 125.292000us
Average time per ec521 evp_set_peer call: 760.357000us
Average time per x25519 evp_set_peer call: 0.957000us
$ ./evp_setpeer -k all -t 2
[dh] 4262.306000us
[ec256] 125.517000us
[ec521] 753.860000us
[x25519] 1.373000us
$ ./evp_setpeer -k x25519 -t 2
[x25519] 1.439600us
```

OpenSSL 1.1.1:
```console
$ ./evp_setpeer -k all -t 2
[dh] 0.394000us
[ec256] 2.347000us
[ec521] 3.217000us
[x25519] 0.103000us
```

From some initial test results, it looks like DH is *much* slower than other key types. Also, OpenSSL 1.1.1 is much faster than OpenSSL 3.3.3, probably because it doesn't validate the peer as suggested in https://github.com/openssl/openssl/issues/21833.

Resolves issue: https://github.com/openssl/project/issues/646
Original PR: https://github.com/openssl/tools/pull/211