@andrewkdinh 2024-08-16
@t8m I tested the results on ToT main, and these are the results. Looks about the same

```console
$ ./evp_setpeer -k all -t 2
[dh] 4254.132400us
[ec256] 123.962700us
[ec521] 758.470000us
[x25519] 1.320900us
```