@t8m 2024-08-16
I believe you are actually using a system libcrypto or something else. The results do not correspond with what I see:

openssl-3.3 branch:
```
$ ./evp_setpeer -k all 1
Average time per dh evp_set_peer call: 1526.177200us
Average time per ec256 evp_set_peer call: 43.988700us
Average time per ec521 evp_set_peer call: 233.551700us
Average time per x25519 evp_set_peer call: 0.334500us
```
master branch:
```
$ ./evp_setpeer -k all 1
Average time per dh evp_set_peer call: 1.787400us
Average time per ec256 evp_set_peer call: 6.231100us
Average time per ec521 evp_set_peer call: 15.156100us
Average time per x25519 evp_set_peer call: 0.694100us
```

These numbers on master reflect the changes of the DH and ECDH key validation to use the quick test for parameters where it is sufficient (i.e., known safe primes such as the FFDHE primes and EC curves with cofactor 1).

So yeah, this call is not _free_ as on 1.1.1 but IMO acceptable.