@t8m 2024-08-16
@andrewkdinh I strongly suggest uncommenting this Makefile line https://github.com/openssl/perftools/blob/main/source/Makefile#L14 when you are building perftools against local builds of openssl.