@andrewkdinh 2024-08-16
@t8m Ah ok, yes it looks like it was because I missed uncommenting that line, and my results look more similar to your results now. Will make sure to do that for future perf results I gather. Thanks!